window: split resize and page-switching into sibling TUs
