+++ panel-plugin/window-geometry.h
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2026 Matteo Bonanomi
+ *
+ * This library is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#ifndef WHISKERMENU_WINDOW_GEOMETRY_H
+#define WHISKERMENU_WINDOW_GEOMETRY_H
+
+/* window-geometry: private translation unit hosting Window's resize and
+ * size-persistence helpers (resize/resize_start/resize_end/set_size and the
+ * scrollbar fit check). Positioning logic (center_window, move_window) and
+ * focus/grab handling intentionally stay in window.cpp.
+ */
+
+#endif // WHISKERMENU_WINDOW_GEOMETRY_H
