+++ panel-plugin/window-pages.h
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2026 Matteo Bonanomi
+ *
+ * This library is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#ifndef WHISKERMENU_WINDOW_PAGES_H
+#define WHISKERMENU_WINDOW_PAGES_H
+
+/* window-pages: private translation unit hosting Window's page-switching
+ * logic. It owns the methods that change which child of m_panels_stack is
+ * visible and which sidebar/category button is active (favorites/recent/
+ * applications toggles, default-button reset, Apps↔Places mode switch, the
+ * search-driven page swap, and active-page/active-category lookup).
+ * Geometry, focus/grab, and positioning logic remain in their own units.
+ */
+
+#endif // WHISKERMENU_WINDOW_PAGES_H
