+++ panel-plugin/window.cpp
@@ -747,30 +747,6 @@ WhiskerMenu::Window::~Window()
 
 //-----------------------------------------------------------------------------
 
-Page* WhiskerMenu::Window::get_active_page()
-{
-	Page* page = nullptr;
-	if (g_strcmp0(gtk_stack_get_visible_child_name(m_panels_stack), "search") == 0)
-	{
-		page = m_search_results;
-	}
-	else if (m_favorites->get_button()->get_active())
-	{
-		page = m_favorites;
-	}
-	else if (m_recent->get_button()->get_active())
-	{
-		page = m_recent;
-	}
-	else
-	{
-		page = m_applications;
-	}
-	return page;
-}
-
-//-----------------------------------------------------------------------------
-
 void WhiskerMenu::Window::hide(bool lost_focus)
 {
 	// Persist Places last-mode so the next open can resume in the same mode.
@@ -1103,58 +1079,6 @@ void WhiskerMenu::Window::show(const Position position)
 
 //-----------------------------------------------------------------------------
 
-void WhiskerMenu::Window::resize(int delta_x, int delta_y, int delta_width, int delta_height)
-{
-	if (set_size(m_geometry.width + delta_width, m_geometry.height + delta_height))
-	{
-		check_scrollbar_needed();
-	}
-
-	if (delta_x || delta_y)
-	{
-		m_geometry.x += delta_x;
-		m_geometry.y += delta_y;
-		move_window();
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::resize_start()
-{
-	m_resizing = true;
-	set_child_has_focus();
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::resize_end()
-{
-	// Store new size (never persist fullscreen dimensions as the normal menu size)
-	if (g_strcmp0(m_settings->layout_mode, "fullscreen") != 0)
-	{
-		m_settings->menu_width = m_geometry.width;
-		m_settings->menu_height = m_geometry.height;
-	}
-
-	// Move window back to panel button or center of screen
-	if (m_position == PositionAtButton)
-	{
-		m_plugin->get_menu_position(&m_geometry.x, &m_geometry.y);
-	}
-	else if (m_position == PositionAtCenter)
-	{
-		center_window();
-	}
-	move_window();
-
-	// Allow menu to hide
-	m_resizing = false;
-	m_child_has_focus = false;
-}
-
-//-----------------------------------------------------------------------------
-
 void WhiskerMenu::Window::set_child_has_focus()
 {
 	m_child_has_focus = true;
@@ -1237,27 +1161,6 @@ void WhiskerMenu::Window::unset_items()
 	m_search_results->unset_menu_items();
 	m_favorites->unset_menu_items();
 	m_recent->unset_menu_items();
-}
-
-//-----------------------------------------------------------------------------
-
-GtkWidget* WhiskerMenu::Window::get_active_category_button()
-{
-	GtkWidget* widget = m_default_button->get_widget();
-
-	GList* children = gtk_container_get_children(GTK_CONTAINER(m_category_buttons));
-	for (GList* li = children; li; li = li->next)
-	{
-		GtkToggleButton* button = GTK_TOGGLE_BUTTON(li->data);
-		if (button && gtk_toggle_button_get_active(button))
-		{
-			widget = GTK_WIDGET(button);
-			break;
-		}
-	}
-	g_list_free(children);
-
-	return widget;
 }
 
 //-----------------------------------------------------------------------------
@@ -1559,55 +1462,6 @@ gboolean WhiskerMenu::Window::on_draw_event(GtkWidget* widget, cairo_t* cr)
 
 //-----------------------------------------------------------------------------
 
-void WhiskerMenu::Window::check_scrollbar_needed()
-{
-	// Find height of sidebar and buttons
-	int buttons_height = 0;
-	gtk_widget_get_preferred_height(GTK_WIDGET(m_category_buttons), nullptr, &buttons_height);
-
-	int sidebar_height = 0;
-	gtk_widget_get_preferred_height(GTK_WIDGET(m_sidebar), nullptr, &sidebar_height);
-
-	// Always show scrollbar if sidebar is shorter than buttons
-	if (sidebar_height >= buttons_height)
-	{
-		gtk_scrolled_window_set_policy(m_sidebar, GTK_POLICY_NEVER, GTK_POLICY_AUTOMATIC);
-	}
-	else
-	{
-		gtk_scrolled_window_set_policy(m_sidebar, GTK_POLICY_NEVER, GTK_POLICY_ALWAYS);
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::favorites_toggled()
-{
-	m_favorites->reset_selection();
-	gtk_stack_set_visible_child_name(m_panels_stack, "favorites");
-	gtk_widget_grab_focus(GTK_WIDGET(m_search_entry));
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::recent_toggled()
-{
-	m_recent->reset_selection();
-	gtk_stack_set_visible_child_name(m_panels_stack, "recent");
-	gtk_widget_grab_focus(GTK_WIDGET(m_search_entry));
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::category_toggled()
-{
-	m_applications->reset_selection();
-	gtk_stack_set_visible_child_name(m_panels_stack, "applications");
-	gtk_widget_grab_focus(GTK_WIDGET(m_search_entry));
-}
-
-//-----------------------------------------------------------------------------
-
 void WhiskerMenu::Window::center_window()
 {
 	m_geometry.x = (m_monitor.width - m_geometry.width) / 2;
@@ -1649,200 +1503,6 @@ void WhiskerMenu::Window::move_window()
 	{
 		gtk_window_move(m_window, m_geometry.x, m_geometry.y);
 	}
-}
-
-//-----------------------------------------------------------------------------
-
-bool WhiskerMenu::Window::set_size(int width, int height)
-{
-	bool resized = false;
-	width = CLAMP(width, 10, m_monitor.width);
-	height = CLAMP(height, 10, m_monitor.height);
-	if ((m_geometry.width != width) || (m_geometry.height != height))
-	{
-		m_geometry.width = width;
-		m_geometry.height = height;
-		gtk_widget_set_size_request(GTK_WIDGET(m_window), m_geometry.width, m_geometry.height);
-		gtk_window_resize(m_window, 1, 1);
-		resized = true;
-	}
-	return resized;
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::reset_default_button()
-{
-	switch (m_settings->default_category)
-	{
-	case Settings::CategoryRecent:
-		m_default_button = m_recent->get_button();
-		gtk_box_reorder_child(m_category_buttons, m_recent->get_button()->get_widget(), 0);
-		gtk_box_reorder_child(m_category_buttons, m_favorites->get_button()->get_widget(), 1);
-		gtk_box_reorder_child(m_category_buttons, m_applications->get_button()->get_widget(), 2);
-		break;
-
-	case Settings::CategoryAll:
-		m_default_button = m_applications->get_button();
-		gtk_box_reorder_child(m_category_buttons, m_applications->get_button()->get_widget(), 0);
-		gtk_box_reorder_child(m_category_buttons, m_favorites->get_button()->get_widget(), 1);
-		gtk_box_reorder_child(m_category_buttons, m_recent->get_button()->get_widget(), 2);
-		break;
-
-	default:
-		m_default_button = m_favorites->get_button();
-		gtk_box_reorder_child(m_category_buttons, m_favorites->get_button()->get_widget(), 0);
-		gtk_box_reorder_child(m_category_buttons, m_recent->get_button()->get_widget(), 1);
-		gtk_box_reorder_child(m_category_buttons, m_applications->get_button()->get_widget(), 2);
-		break;
-	}
-
-	// NOTE: the three reorders above push m_mode_selector_box (packed first
-	// in the ctor) past the category buttons. Restore it to the top so the
-	// Apps/Places switch stays above Recently Used/Favorites/All Applications.
-	if (m_mode_selector_box)
-	{
-		gtk_box_reorder_child(m_category_buttons, GTK_WIDGET(m_mode_selector_box), 0);
-	}
-	if (m_mode_selector_separator)
-	{
-		gtk_box_reorder_child(m_category_buttons, m_mode_selector_separator, 1);
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::show_favorites()
-{
-	// Switch to favorites panel
-	m_favorites->get_button()->set_active(true);
-
-	// Clear search entry
-	gtk_entry_set_text(m_search_entry, "");
-	gtk_widget_grab_focus(GTK_WIDGET(m_search_entry));
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::show_default_page()
-{
-	// Switch to favorites panel
-	m_default_button->set_active(true);
-
-	// Clear search entry
-	gtk_entry_set_text(m_search_entry, "");
-	gtk_widget_grab_focus(GTK_WIDGET(m_search_entry));
-}
-
-//-----------------------------------------------------------------------------
-
-void WhiskerMenu::Window::search()
-{
-	// Fetch search string
-	const gchar* text = gtk_entry_get_text(m_search_entry);
-	if (xfce_str_is_empty(text))
-	{
-		text = nullptr;
-	}
-
-	if (m_places_active)
-	{
-		// Places mode: stay on the places panel; filter the active section.
-		gtk_stack_set_visible_child_name(m_panels_stack, "places");
-		m_places->set_filter(text);
-		return;
-	}
-
-	if (text)
-	{
-		// Switch the applications area to search results; the sidebar stays visible.
-		gtk_stack_set_visible_child_name(m_panels_stack, "search");
-	}
-	else
-	{
-		// Restore the panel that matches the currently active category button.
-		if (m_favorites->get_button()->get_active())
-			gtk_stack_set_visible_child_name(m_panels_stack, "favorites");
-		else if (m_recent->get_button()->get_active())
-			gtk_stack_set_visible_child_name(m_panels_stack, "recent");
-		else
-			gtk_stack_set_visible_child_name(m_panels_stack, "applications");
-	}
-
-	// Apply filter
-	m_search_results->set_filter(text);
-}
-
-//-----------------------------------------------------------------------------
-
-/* switch_mode:
- * @to_places: true to enter Places mode; false to return to Apps.
- *
- * Toggles the Apps/Places selector visuals, hides/shows the appropriate
- * sidebar buttons, and updates the search-entry placeholder. Re-entrancy is
- * guarded by m_mode_switch_in_progress so the two toggle-button "toggled"
- * signals do not loop.
- */
-void WhiskerMenu::Window::switch_mode(bool to_places)
-{
-	if (m_mode_switch_in_progress)
-	{
-		return;
-	}
-	m_mode_switch_in_progress = true;
-
-	m_places_active = to_places;
-	gtk_toggle_button_set_active(m_mode_btn_apps,   !to_places);
-	gtk_toggle_button_set_active(m_mode_btn_places,  to_places);
-
-	const bool history_visible = m_settings->places_history_enabled;
-	const bool fav_visible     = m_settings->places_favourites_enabled;
-
-	gtk_widget_set_visible(m_favorites->get_button()->get_widget(),       !to_places);
-	gtk_widget_set_visible(m_recent->get_button()->get_widget(),
-			!to_places && m_settings->recent_items_max);
-	gtk_widget_set_visible(m_applications->get_button()->get_widget(),    !to_places);
-
-	gtk_widget_set_visible(m_places_home_btn->get_widget(),     to_places);
-	gtk_widget_set_visible(m_places_history_btn->get_widget(),  to_places && history_visible);
-	gtk_widget_set_visible(m_places_fav_btn->get_widget(),      to_places && fav_visible);
-
-	// Hide app categories (Accessories, Development, ...) in Places mode.
-	for (GtkWidget* w : m_app_category_widgets)
-	{
-		gtk_widget_set_visible(w, !to_places);
-	}
-
-	gtk_entry_set_text(m_search_entry, "");
-	gtk_entry_set_placeholder_text(m_search_entry,
-			to_places ? _("Search places\xe2\x80\xa6")
-			          : _("Search applications\xe2\x80\xa6"));
-
-	if (to_places)
-	{
-		m_places_home_btn->set_active(true);
-		m_places->set_active_section(m_places->get_home_section());
-		gtk_stack_set_visible_child_name(m_panels_stack, "places");
-	}
-	else
-	{
-		// NOTE: Apps and Places buttons are separate radio groups. When Places
-		// was entered, the default Apps button was never deactivated, so
-		// set_active(true) in show_default_page() is a GTK no-op and the
-		// "toggled" handler never fires. Switch the stack explicitly first.
-		const char* page = "favorites";
-		switch (m_settings->default_category)
-		{
-		case Settings::CategoryRecent: page = "recent";       break;
-		case Settings::CategoryAll:    page = "applications"; break;
-		default: break;
-		}
-		gtk_stack_set_visible_child_name(m_panels_stack, page);
-		show_default_page();
-	}
-
-	gtk_widget_grab_focus(GTK_WIDGET(m_search_entry));
-	m_mode_switch_in_progress = false;
 }
 
 //-----------------------------------------------------------------------------
